added styling for all, active and completed button when the button is active and code for functionality of buttons when clicked to show the appropriate tasks

+++ src/components/Todos.jsx
@@ -4,6 +4,7 @@ import {useState} from 'react';
 function Todos(){
     const [items, setItems] = useState([])
     const [newItemText, setNewItemText] =  useState('')
+    const [currentTab, setCurrentTab] =  useState('All')
     const addItem = async () => {
         
         setItems([...items, {txt: newItemText, completed: false}])
@@ -16,6 +17,7 @@ function Todos(){
         setNewItemText(event.currentTarget.value)
 
     };
+
     const toggleCompleted = async (event,idx)=>{
         console.log(event.target.checked)
         let newItemsList = []
@@ -48,15 +50,35 @@ function Todos(){
                 <button className="add-btn" onClick={addItem}>Add</button>
             </div>
 
+            <p>{JSON.stringify(items)}</p>
+            <p>{currentTab}</p>
+
             <div className="todo-list-container">
                 <div className="todo-list">
                     <ul className="todo-ul">
                         {items && items.length > 0 && items.map((item, idx) => {
-                            return <li className="task-li" key={idx}>
+                            // is the current item in the loop completed AND is the current tab == completed, if yes show the item
+                            if(item.completed && currentTab == 'Completed'){
+                                return <li className="task-li" key={idx}>
                             <input type="checkbox" className="checkbox" onChange={(event)=> toggleCompleted(event, idx)}/>
                             <span>{item.txt}</span>
                             <i class="fa-regular fa-trash-can" onClick={(event)=> deleteItem(event, idx)} style={{float:"right", color:"red", marginRight:"10px"}}></i>
                         </li>
+                            } else if(!item.completed && currentTab == 'Active') {
+                                return <li className="task-li" key={idx}>
+                            <input type="checkbox" className="checkbox" onChange={(event)=> toggleCompleted(event, idx)}/>
+                            <span>{item.txt}</span>
+                            <i class="fa-regular fa-trash-can" onClick={(event)=> deleteItem(event, idx)} style={{float:"right", color:"red", marginRight:"10px"}}></i>
+                        </li>
+
+                            } else if(currentTab == 'All'){
+                                return <li className="task-li" key={idx}>
+                            <input type="checkbox" className="checkbox" onChange={(event)=> toggleCompleted(event, idx)}/>
+                            <span>{item.txt}</span>
+                            <i class="fa-regular fa-trash-can" onClick={(event)=> deleteItem(event, idx)} style={{float:"right", color:"red", marginRight:"10px"}}></i>
+                        </li>
+                            }
+                        
                         })} {/*when item exist and when there is more than one item then loop over each items grab one item at a time and for each item return this*/}
                         
                         {/*<li className="task-li">
@@ -77,9 +99,9 @@ function Todos(){
 
             <div className="show-container">
                 <div>
-                    <button className="all-btn active">All</button>
-                    <button className="active-btn">Active</button>
-                    <button className="complete-btn">Complete</button>
+                    <button onClick={() => setCurrentTab('All')} className={currentTab=='All' ?  "all-btn-active" : "all-btn"}>All</button>
+                    <button onClick={() => setCurrentTab('Active')} className={currentTab=='Active' ?  "active-btn-active" : "active-btn"}>Active</button>
+                    <button onClick={() => setCurrentTab('Completed')} className={currentTab=='Completed' ?  "complete-btn-active" : "complete-btn"}>Complete</button>
                 </div>
